Truncate warning messages, and make sure only 10 blocks are returned in an error message
